+++ wallet/service/v2/connections/store/v1/file_store.go
@@ -1,26 +1,42 @@
 package v1
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"os"
+	"sync"
 	"time"
 
 	vgfs "code.vegaprotocol.io/vega/libs/fs"
+	vgjob "code.vegaprotocol.io/vega/libs/job"
 	"code.vegaprotocol.io/vega/paths"
 	"code.vegaprotocol.io/vega/wallet/api"
 	"code.vegaprotocol.io/vega/wallet/service/v2/connections"
+	"github.com/fsnotify/fsnotify"
 )
 
 var ErrStoreNotInitialized = errors.New("the tokens store has not been initialized")
 
-type Store struct {
+type FileStore struct {
 	tokensFilePath string
-	passphrase     string
+
+	passphrase string
+
+	// jobRunner is used to start and stop the file watcher routines.
+	jobRunner *vgjob.Runner
+
+	// listeners are callback functions to be called when a change occurs on
+	// the tokens.
+	listeners []func(context.Context, ...connections.TokenDescription)
+	mu        sync.Mutex
 }
 
-func (s *Store) TokenExists(token connections.Token) (bool, error) {
-	tokens, err := s.readFile()
+func (s *FileStore) TokenExists(token connections.Token) (bool, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	tokens, err := s.readTokensFile()
 	if err != nil {
 		return false, err
 	}
@@ -34,15 +50,18 @@ func (s *Store) TokenExists(token connections.Token) (bool, error) {
 	return false, nil
 }
 
-func (s *Store) ListTokens() ([]connections.TokenSummary, error) {
-	tokens, err := s.readFile()
+func (s *FileStore) ListTokens() ([]connections.TokenSummary, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	tokensFile, err := s.readTokensFile()
 	if err != nil {
 		return nil, err
 	}
 
-	summaries := make([]connections.TokenSummary, 0, len(tokens.Tokens))
+	summaries := make([]connections.TokenSummary, 0, len(tokensFile.Tokens))
 
-	for _, tokenInfo := range tokens.Tokens {
+	for _, tokenInfo := range tokensFile.Tokens {
 		token, err := connections.AsToken(tokenInfo.Token)
 		if err != nil {
 			return nil, fmt.Errorf("token %q is not a valid token: %w", token, err)
@@ -58,8 +77,11 @@ func (s *Store) ListTokens() ([]connections.TokenSummary, error) {
 	return summaries, nil
 }
 
-func (s *Store) DescribeToken(token connections.Token) (connections.TokenDescription, error) {
-	tokens, err := s.readFile()
+func (s *FileStore) DescribeToken(token connections.Token) (connections.TokenDescription, error) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	tokens, err := s.readTokensFile()
 	if err != nil {
 		return connections.TokenDescription{}, err
 	}
@@ -82,27 +104,33 @@ func (s *Store) DescribeToken(token connections.Token) (connections.TokenDescrip
 	return connections.TokenDescription{}, ErrTokenDoesNotExist
 }
 
-func (s *Store) SaveToken(token connections.TokenDescription) error {
-	tokens, err := s.readFile()
+func (s *FileStore) SaveToken(token connections.TokenDescription) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	tokensFile, err := s.readTokensFile()
 	if err != nil {
 		return err
 	}
 
-	tokens.Resources.Wallets[token.Wallet.Name] = token.Wallet.Passphrase
+	tokensFile.Resources.Wallets[token.Wallet.Name] = token.Wallet.Passphrase
 
-	tokens.Tokens = append(tokens.Tokens, tokenContent{
+	tokensFile.Tokens = append(tokensFile.Tokens, tokenContent{
 		Token:          token.Token.String(),
-		CreationDate:   time.Now(),
+		CreationDate:   token.CreationDate,
 		Description:    token.Description,
 		Wallet:         token.Wallet.Name,
 		ExpirationDate: token.ExpirationDate,
 	})
 
-	return s.writeFile(tokens)
+	return s.writeTokensFile(tokensFile)
 }
 
-func (s *Store) DeleteToken(token connections.Token) error {
-	tokens, err := s.readFile()
+func (s *FileStore) DeleteToken(token connections.Token) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	tokens, err := s.readTokensFile()
 	if err != nil {
 		return err
 	}
@@ -125,11 +153,31 @@ func (s *Store) DeleteToken(token connections.Token) error {
 		}
 	}
 
-	return s.writeFile(tokens)
+	return s.writeTokensFile(tokens)
 }
 
-func (s *Store) readFile() (tokensFile, error) {
+func (s *FileStore) OnUpdate(callbackFn func(ctx context.Context, tokens ...connections.TokenDescription)) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	s.listeners = append(s.listeners, callbackFn)
+}
+
+func (s *FileStore) Close() {
+	if s.jobRunner != nil {
+		s.jobRunner.StopAllJobs()
+	}
+}
+
+func (s *FileStore) readTokensFile() (tokensFile, error) {
 	tokens := tokensFile{}
+
+	exists, err := vgfs.FileExists(s.tokensFilePath)
+	if err != nil {
+		return tokensFile{}, fmt.Errorf("could not verify the existence of the tokens file: %w", err)
+	} else if !exists {
+		return defaultTokensFileContent(), nil
+	}
 
 	if err := paths.ReadEncryptedFile(s.tokensFilePath, s.passphrase, &tokens); err != nil {
 		if err.Error() == "couldn't decrypt content: cipher: message authentication failed" {
@@ -149,7 +197,7 @@ func (s *Store) readFile() (tokensFile, error) {
 	return tokens, nil
 }
 
-func (s *Store) writeFile(tokens tokensFile) error {
+func (s *FileStore) writeTokensFile(tokens tokensFile) error {
 	if err := paths.WriteEncryptedFile(s.tokensFilePath, s.passphrase, tokens); err != nil {
 		return fmt.Errorf("couldn't write the file %s: %w", s.tokensFilePath, err)
 	}
@@ -157,7 +205,7 @@ func (s *Store) writeFile(tokens tokensFile) error {
 	return nil
 }
 
-func (s *Store) wipeOut() error {
+func (s *FileStore) wipeOut() error {
 	exists, err := vgfs.FileExists(s.tokensFilePath)
 	if err != nil {
 		return fmt.Errorf("could not verify the existence of the tokens file: %w", err)
@@ -172,19 +220,146 @@ func (s *Store) wipeOut() error {
 	return nil
 }
 
-func (s *Store) initDefault() error {
-	return s.writeFile(defaultTokensFileContent())
+func (s *FileStore) initDefault() error {
+	return s.writeTokensFile(defaultTokensFileContent())
 }
 
-func LoadStore(p paths.Paths, passphrase string) (*Store, error) {
+func (s *FileStore) startFileWatcher() error {
+	watcher, err := fsnotify.NewWatcher()
+	if err != nil {
+		return fmt.Errorf("could not start the token store watcher: %w", err)
+	}
+
+	s.jobRunner = vgjob.NewRunner(context.Background())
+
+	s.jobRunner.Go(func(ctx context.Context) {
+		s.watchFile(ctx, watcher)
+	})
+
+	if err := watcher.Add(s.tokensFilePath); err != nil {
+		return fmt.Errorf("could not start watching the token file: %w", err)
+	}
+
+	return nil
+}
+
+func (s *FileStore) watchFile(ctx context.Context, watcher *fsnotify.Watcher) {
+	defer func() {
+		_ = watcher.Close()
+	}()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case event, ok := <-watcher.Events:
+			if !ok {
+				return
+			}
+			s.broadcastFileChanges(ctx, watcher, event)
+		case _, ok := <-watcher.Errors:
+			if !ok {
+				return
+			}
+			// Something went wrong, but tons of thing can go wrong on a file
+			// system, and there is nothing we can do about that. Let's ignore it.
+		}
+	}
+}
+
+func (s *FileStore) broadcastFileChanges(ctx context.Context, watcher *fsnotify.Watcher, event fsnotify.Event) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
+	if event.Op == fsnotify.Chmod {
+		// If this is solely a CHMOD, we do not trigger an update.
+		return
+	}
+
+	if event.Has(fsnotify.Remove) {
+		_ = watcher.Remove(s.tokensFilePath)
+		exists, err := vgfs.FileExists(s.tokensFilePath)
+		if err != nil {
+			return
+		}
+		if !exists {
+			// The file could have been re-created before we acquire the
+			// lock.
+			_ = s.initDefault()
+		}
+		_ = watcher.Add(s.tokensFilePath)
+	}
+
+	// Let's wait a bit so any write actions in progress have a chance to finish.
+	// This is far from being resilient, but it can help.
+	time.Sleep(100 * time.Millisecond)
+
+	tokenDescriptions, err := s.readFileAsTokenDescriptions()
+	if err != nil {
+		// This can be the result of concurrent modification on the token file.
+		// The best thing to do is to carry on and ignore the changes.
+		return
+	}
+
+	for _, listener := range s.listeners {
+		listener(ctx, tokenDescriptions...)
+	}
+}
+
+func (s *FileStore) readFileAsTokenDescriptions() ([]connections.TokenDescription, error) {
+	exists, err := vgfs.FileExists(s.tokensFilePath)
+	if err != nil {
+		return nil, fmt.Errorf("could not verify the token file existence: %w", err)
+	}
+	if !exists {
+		// The file got deleted.
+		//
+		// This can the result of a "desperate" action to kill all long-living
+		// connections, using the `api-token init --force` command.
+		//
+		// As a result, we return an empty list of tokens, meaning, all tokens
+		// should be considered invalid from now on.
+		return nil, nil
+	}
+
+	tokensFile, err := s.readTokensFile()
+	if err != nil {
+		return nil, fmt.Errorf("could not read the token file: %w", err)
+	}
+
+	tokenDescriptions := make([]connections.TokenDescription, 0, len(tokensFile.Tokens))
+
+	for _, tokenInfo := range tokensFile.Tokens {
+		token, err := connections.AsToken(tokenInfo.Token)
+		if err != nil {
+			// It's all or nothing.
+			return nil, fmt.Errorf("the token %q could not be parse: %w", tokenInfo.Token, err)
+		}
+		tokenDescriptions = append(tokenDescriptions, connections.TokenDescription{
+			Description:    tokenInfo.Description,
+			CreationDate:   tokenInfo.CreationDate,
+			ExpirationDate: tokenInfo.ExpirationDate,
+			Token:          token,
+			Wallet: connections.WalletCredentials{
+				Name:       tokenInfo.Wallet,
+				Passphrase: tokensFile.Resources.Wallets[tokenInfo.Wallet],
+			},
+		})
+	}
+
+	return tokenDescriptions, nil
+}
+
+func InitialiseStore(p paths.Paths, passphrase string) (*FileStore, error) {
 	tokensFilePath, err := p.CreateDataPathFor(paths.WalletServiceTokensDataFile)
 	if err != nil {
 		return nil, fmt.Errorf("couldn't get data path for %s: %w", paths.WalletServicePublicRSAKeyDataFile, err)
 	}
 
-	store := &Store{
+	store := &FileStore{
 		tokensFilePath: tokensFilePath,
 		passphrase:     passphrase,
+		listeners:      []func(context.Context, ...connections.TokenDescription){},
 	}
 
 	exists, err := vgfs.FileExists(tokensFilePath)
@@ -192,33 +367,28 @@ func LoadStore(p paths.Paths, passphrase string) (*Store, error) {
 		return nil, ErrStoreNotInitialized
 	}
 
-	if _, err := store.readFile(); err != nil {
+	if _, err := store.readTokensFile(); err != nil {
+		return nil, err
+	}
+
+	if err := store.startFileWatcher(); err != nil {
+		store.Close()
 		return nil, err
 	}
 
 	return store, nil
 }
 
-func IsStoreInitialized(p paths.Paths) (bool, error) {
-	tokensFilePath, err := tokensFilePath(p)
-	if err != nil {
-		return false, err
-	}
-
-	exists, err := vgfs.FileExists(tokensFilePath)
-
-	return err == nil && exists, nil
-}
-
-func InitializeStore(p paths.Paths, passphrase string) (*Store, error) {
+func ReinitialiseStore(p paths.Paths, passphrase string) (*FileStore, error) {
 	tokensFilePath, err := tokensFilePath(p)
 	if err != nil {
 		return nil, err
 	}
 
-	store := &Store{
+	store := &FileStore{
 		tokensFilePath: tokensFilePath,
 		passphrase:     passphrase,
+		listeners:      []func(context.Context, ...connections.TokenDescription){},
 	}
 
 	if err := store.wipeOut(); err != nil {
@@ -229,7 +399,23 @@ func InitializeStore(p paths.Paths, passphrase string) (*Store, error) {
 		return nil, err
 	}
 
+	if err := store.startFileWatcher(); err != nil {
+		store.Close()
+		return nil, err
+	}
+
 	return store, nil
+}
+
+func IsStoreBootstrapped(p paths.Paths) (bool, error) {
+	tokensFilePath, err := tokensFilePath(p)
+	if err != nil {
+		return false, err
+	}
+
+	exists, err := vgfs.FileExists(tokensFilePath)
+
+	return err == nil && exists, nil
 }
 
 func tokensFilePath(p paths.Paths) (string, error) {